:truck:move: 잡담게시글 service 파일 이동 (#1)
