+++ src/main/java/com/example/backend/community/CommunityController.java
@@ -2,13 +2,13 @@ package com.example.backend.community;
 
 import com.example.backend.community.domain.CommunityBoard;
 import com.example.backend.community.dto.CommunityBoardDto;
+import com.example.backend.community.service.CommunityBoardService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
-import org.springframework.web.multipart.MultipartFile;
 
 import java.time.LocalDateTime;
